Replace int with size_t

+++ src/algorithms/sort/quick sort/main.c
@@ -6,8 +6,8 @@
 
 int main()
 {
-    int array[] = {7, 1, 326, 832, 225, 32323, 45, 1, 36, 43534, 34534, 2};
+    size_t array[] = {7, 1, 326, 832, 225, 32323, 45, 1, 36, 43534, 34534, 2};
     quick_sort(array, 0, LENGTH - 1);
     for (int i = 0; i < LENGTH; ++i)
-        printf("%d\n", array[i]);
+        printf("%zd\n", array[i]);
 }
